#38
- preset/isomorphic: HttpClientLogging переименован в AxiosLogging (major)
- preset/bun: FetchLogging перенесен в preset/isomorphic (patch)
