+++ src/preset/bun/bun/utils/index.ts
@@ -1,76 +1,12 @@
 /* eslint-disable require-jsdoc, jsdoc/require-jsdoc */
 import { DoneLogData, FailLogData, FetchUtil, LogData, LogHandler } from '../../../../http';
-import { Breadcrumb, DetailedError, Logger } from '../../../../log';
+import { DetailedError, Logger } from '../../../../log';
 import { toMilliseconds } from '../../../../utils';
 
 export function getClientIp(request: Request): string | null {
   const headerValue = request.headers.get('x-client-ip') || request.headers.get('x-forwarded-for');
 
   return headerValue;
-}
-
-/**
- * Обработчик логирования внешних http-запросов.
- */
-export class FetchLogging implements LogHandler {
-  logger: Logger;
-  disabled: boolean;
-
-  constructor(logger: Logger) {
-    this.logger = logger;
-    this.disabled = false;
-  }
-
-  onRequest({ request }: LogData) {
-    this.logger.info(
-      new Breadcrumb({
-        category: 'http.request',
-        type: 'http',
-        data: {
-          url: FetchUtil.withoutParams(request.url),
-          method: request.method,
-          params: Object.fromEntries(new URL(request.url).searchParams.entries()),
-        },
-        level: 'info',
-      }),
-    );
-  }
-
-  onResponse({ response, request }: DoneLogData) {
-    this.logger.info(
-      new Breadcrumb({
-        category: 'http.response',
-        type: 'http',
-        data: {
-          url: FetchUtil.withoutParams(request.url),
-          method: request.method,
-          status_code: response.status,
-          params: Object.fromEntries(new URL(request.url).searchParams.entries()),
-        },
-        level: response.ok ? 'info' : 'error',
-      }),
-    );
-  }
-
-  onCatch({ error, request }: FailLogData) {
-    this.logger.error(
-      new DetailedError(String(error), {
-        level: 'error',
-        context: [
-          {
-            key: 'Outgoing request details',
-            data: {
-              url: FetchUtil.withoutParams(request.url),
-              method: request.method,
-              headers: request.headers,
-              params: Object.fromEntries(new URL(request.url).searchParams.entries()),
-              // @todo data
-            },
-          },
-        ],
-      }),
-    );
-  }
 }
 
 /**
